requirements.txt
# =============================================================================
# LLM Cybersecurity Course — Unified Requirements
# Author: Badr TAJINI — ECE 2025/2026
# Python: 3.11+ recommended (3.12 max for semgrep compatibility)
# Node.js: 22.x or 24.x LTS for promptfoo
# Install: pip install -r requirements.txt
# =============================================================================

# -----------------------------------------------------------------------------
# Core: Gemini API (google-genai 1.55.0 requires Python ≥3.10)
# Latest: 1.55.0 - https://pypi.org/project/google-genai/
# -----------------------------------------------------------------------------
google-genai>=1.55.0

# -----------------------------------------------------------------------------
# Environment & Config
# -----------------------------------------------------------------------------
python-dotenv>=1.0.1
pyyaml>=6.0.2

# -----------------------------------------------------------------------------
# Validation - LET SEMGREP CONTROL THIS (pins pydantic~=2.8.2)
# Do not specify pydantic version explicitly to avoid conflicts
# -----------------------------------------------------------------------------
# pydantic is installed as semgrep dependency

# -----------------------------------------------------------------------------
# Security Scanners (Lab 3)
# semgrep 1.145.x requires: Python 3.10+, rich~=13.5.2, pydantic~=2.8.2
# checkov 3.2.495 - https://pypi.org/project/checkov/
# -----------------------------------------------------------------------------
semgrep>=1.145.0
checkov>=3.2.495

# -----------------------------------------------------------------------------
# ML/Data (Lab 4 + Project RAG)
# scikit-learn 1.6.1 - https://pypi.org/project/scikit-learn/
# -----------------------------------------------------------------------------
scikit-learn>=1.6.0

# -----------------------------------------------------------------------------
# Output formatting - LET SEMGREP CONTROL THIS (pins rich~=13.5.2)
# -----------------------------------------------------------------------------
# rich is installed as semgrep dependency

# -----------------------------------------------------------------------------
# Testing
# -----------------------------------------------------------------------------
pytest>=8.3.0
pytest-mock>=3.14.0

# -----------------------------------------------------------------------------
# Notebooks (optional, for Jupyter support)
# -----------------------------------------------------------------------------
nbformat>=5.10.0
nbclient>=0.10.0
ipykernel>=6.29.0
